pyproject.toml
[tool.poetry]
name = "python-ml-server"
version = "0.1.0"
description = ""
authors = ["Sean Sun <tseensun@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.96.1"
uvicorn = "^0.22.0"
pydantic = {extras = ["email"], version = "^1.10.9"}
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.16"
alembic = "^1.11.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
types-python-jose = "^3.3.4.7"
types-passlib = "^1.7.7.12"
openai-clip = "^1.0.1"
torch = "^2.0.1"
torchvision = "^0.15.2"
pillow = "^9.5.0"
types-pillow = "^9.5.0.4"
tenacity = "^8.2.2"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
sqlalchemy-stubs = "^0.4"

[tool.poe.tasks]
install-ltt = "python3 -m pip install light-the-torch"
run-ltt = "python3 -m light_the_torch install --upgrade torch torchaudio torchvision"
autoinstall-torch-cuda = ["install-ltt", "run-ltt"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


